feat: add a route point switch to the form
