+++ src/view/route-point-view.js
@@ -1,204 +1,30 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import {humanizeTaskDueDate, showTripDuration, calculateTripDuration, showFullDate, showFullDateTime, showNewPointDate} from '../utils.js';
-
-function createAddNewPointWithout(type, offers, startDate, endDate, basePrice) {
-  return `<form class="event event--edit hidden" action="#" method="post">
-            <header class="event__header">
-              <div class="event__type-wrapper">
-                <label class="event__type  event__type-btn" for="event-type-toggle-1">
-                  <span class="visually-hidden">Choose event type</span>
-                  <img class="event__type-icon" width="17" height="17" src="img/icons/${type}.png" alt="Event type icon">
-                </label>
-                <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
-
-                <div class="event__type-list">
-                  <fieldset class="event__type-group">
-                    <legend class="visually-hidden">Event type</legend>
-
-                    <div class="event__type-item">
-                      <input id="event-type-taxi-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="taxi">
-                      <label class="event__type-label  event__type-label--taxi" for="event-type-taxi-1">Taxi</label>
-                    </div>
-
-                    <div class="event__type-item">
-                      <input id="event-type-bus-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="bus">
-                      <label class="event__type-label  event__type-label--bus" for="event-type-bus-1">Bus</label>
-                    </div>
-
-                    <div class="event__type-item">
-                      <input id="event-type-train-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="train">
-                      <label class="event__type-label  event__type-label--train" for="event-type-train-1">Train</label>
-                    </div>
-
-                    <div class="event__type-item">
-                      <input id="event-type-ship-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="ship">
-                      <label class="event__type-label  event__type-label--ship" for="event-type-ship-1">Ship</label>
-                    </div>
-
-                    <div class="event__type-item">
-                      <input id="event-type-drive-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="drive">
-                      <label class="event__type-label  event__type-label--drive" for="event-type-drive-1">Drive</label>
-                    </div>
-
-                    <div class="event__type-item">
-                      <input id="event-type-flight-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="flight" checked>
-                      <label class="event__type-label  event__type-label--flight" for="event-type-flight-1">Flight</label>
-                    </div>
-
-                    <div class="event__type-item">
-                      <input id="event-type-check-in-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="check-in">
-                      <label class="event__type-label  event__type-label--check-in" for="event-type-check-in-1">Check-in</label>
-                    </div>
-
-                    <div class="event__type-item">
-                      <input id="event-type-sightseeing-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="sightseeing">
-                      <label class="event__type-label  event__type-label--sightseeing" for="event-type-sightseeing-1">Sightseeing</label>
-                    </div>
-
-                    <div class="event__type-item">
-                      <input id="event-type-restaurant-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="restaurant">
-                      <label class="event__type-label  event__type-label--restaurant" for="event-type-restaurant-1">Restaurant</label>
-                    </div>
-                  </fieldset>
-                </div>
-              </div>
-
-              <div class="event__field-group  event__field-group--destination">
-                <label class="event__label  event__type-output" for="event-destination-1">
-                  ${type}
-                </label>
-                <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="" list="destination-list-1">
-                <datalist id="destination-list-1">
-                  <option value="Amsterdam"></option>
-                  <option value="Geneva"></option>
-                  <option value="Chamonix"></option>
-                </datalist>
-              </div>
-
-              <div class="event__field-group  event__field-group--time">
-                <label class="visually-hidden" for="event-start-time-1">From</label>
-                <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="${showNewPointDate(startDate)}">
-                &mdash;
-                <label class="visually-hidden" for="event-end-time-1">To</label>
-                <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="${showNewPointDate(endDate)}">
-              </div>
-
-              <div class="event__field-group  event__field-group--price">
-                <label class="event__label" for="event-price-1">
-                  <span class="visually-hidden">Price</span>
-                  &euro; ${basePrice}
-                </label>
-                <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="">
-              </div>
-
-              <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
-              <button class="event__reset-btn" type="reset">Cancel</button>
-            </header>
-            <section class="event__details">
 
 
-              <section class="event__section  event__section--offers">
-                <h3 class="event__section-title  event__section-title--offers">Offers</h3>
-                <div class="event__available-offers">
-                  ${createButtonAddServices(offers)}
-                  </div>
-                </div>
-              </section>
-
-
-            </section>
-          </form>`;
-}
-
-function createButtonAddServices(offers) {
-  return offers.map((offer) =>
-    `<div class="event__offer-selector">
-      <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-${offer.id}" type="checkbox" name="event-offer-comfort" checked>
-      <label class="event__offer-label" for="event-offer-comfort-${offer.id}">
-        <span class="event__offer-title">${offer.title}</span>
-        &plus;&euro;&nbsp;
-        <span class="event__offer-price">${offer.price}</span>
-      </label>
-    </div>`
-  ).join('');
-}
-
-function createAddServices(offers) {
-  return offers.map((offer) =>
-    `<li class="event__offer">
-      <span class="event__offer-title">${offer.title}</span>
-      &plus;&euro;&nbsp;
-      <span class="event__offer-price">${offer.price}</span>
-    </li>`
-  ).join('');
-}
-
-function createRoutPointTemplate(task) {
-  // console.log(task);
-  const {destinationDetails, type, offers, startDate, endDate, basePrice} = task;
+function createRoutPointTemplate(task, cbForm, cbBox) {
+  const {type, offers, startDate, endDate, basePrice} = task;
+  const additionalContentForm = typeof cbForm === 'function' ? cbForm(type, offers, startDate, endDate, basePrice) : '';
+  const additionalContentBox = typeof cbBox === 'function' ? cbBox(task) : '';
 
   return `<li class="trip-events__item">
-            <div class="event">
-              <time class="event__date" datetime="${showFullDate(startDate)}">${humanizeTaskDueDate(startDate)}</time>
-              <div class="event__type">
-                <img class="event__type-icon" width="42" height="42" src="img/icons/${type}.png" alt="Event type icon">
-              </div>
-              <h3 class="event__title">${type} ${destinationDetails.name}</h3>
-              <div class="event__schedule">
-                <p class="event__time">
-                  <time class="event__start-time" datetime="${showFullDateTime(startDate)}">${showTripDuration(startDate)}</time>
-                  &mdash;
-                  <time class="event__end-time" datetime="${showFullDateTime(endDate)}">${showTripDuration(endDate)}</time>
-                </p>
-                <p class="event__duration">${calculateTripDuration(startDate, endDate)}</p>
-              </div>
-              <p class="event__price">
-                &euro;&nbsp;<span class="event__price-value">${basePrice}</span>
-              </p>
-              <h4 class="visually-hidden">Offers:</h4>
-              <ul class="event__selected-offers">
-                ${createAddServices(offers)}
-              </ul>
-              <button class="event__favorite-btn  event__favorite-btn--active" type="button">
-                <span class="visually-hidden">Add to favorite</span>
-                <svg class="event__favorite-icon" width="28" height="28" viewBox="0 0 28 28">
-                  <path d="M14 21l-8.22899 4.3262 1.57159-9.1631L.685209 9.67376 9.8855 8.33688 14 0l4.1145 8.33688 9.2003 1.33688-6.6574 6.48934 1.5716 9.1631L14 21z"/>
-                </svg>
-              </button>
-              <button class="event__rollup-btn" type="button">
-                <span class="visually-hidden">Open event</span>
-              </button>
-            </div>
-            ${createAddNewPointWithout(type, offers, startDate, endDate, basePrice)}
+            ${additionalContentBox}
+            ${additionalContentForm}
           </li>`;
 }
 
 export default class RoutPointView extends AbstractView {
   #task = null;
-  #element = null;
-  constructor({task}) {
+  #cbForm = null;
+  #cbBox = null;
+
+  constructor({task, cbForm, cbBox}) {
     super();
     this.#task = task;
-    this.#element = null;
+    this.#cbForm = cbForm;
+    this.#cbBox = cbBox;
   }
 
   get template() {
-    return createRoutPointTemplate(this.#task);
+    return createRoutPointTemplate(this.#task, this.#cbForm, this.#cbBox);
   }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = super.element;
-      const eventRollupButtonElement = this.#element.querySelector('.event__rollup-btn');
-      eventRollupButtonElement.addEventListener('click', this.handleRollupButtonClick);
-    }
-    return this.#element;
-  }
-
-  handleRollupButtonClick = () => {
-    const formElement = this.element.querySelector('.event--edit');
-    const pointElement = this.element.querySelector('.event');
-    formElement.classList.toggle('hidden');
-    pointElement.classList.toggle('hidden');
-  };
 }
